removed "using namespace std;'
This is bad practice:
    https://stackoverflow.com/questions/1452721/why-is-using-namespace-std-considered-bad-practice

Cleaned up the use of std::cin/std::cout
There format around the input/operator needed extra space to be readable.

Removed the std::endl and replaced with "\n".
This is because this is bad practice as it forces an unneeded flush of the stream.

std::swap is not included by default through `iostream` so added appropriate includes.